requirements.txt
google-genai>=0.7.0
opencv-python>=4.8.0
pyaudio>=0.2.11
pillow>=10.0.0
mss>=9.0.0
python-dotenv>=1.0.0
